requirements.txt
# Core machine learning framework
tensorflow>=2.10.0
keras>=2.10.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.6.0
Pillow>=9.0.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.9.0

# Web application framework
gradio>=3.35.0

# Configuration management
pyyaml>=6.0
python-dotenv>=0.20.0

# Utility tools
tqdm>=4.64.0
h5py>=3.7.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Project development
black>=22.0.0
flake8>=5.0.0
pytest>=7.0.0